renamed pump feature sub-folders
