+++ lib/src/config/mock/fake_pumps.dart
@@ -1,4 +1,4 @@
-import 'package:irrigazione_iot/src/features/pumps/domain/pump.dart';
+import 'package:irrigazione_iot/src/features/pumps/model/pump.dart';
 
 const kFakePumps = [
   Pump(
