+++ lib/src/features/pumps/data/pump_details_repository.dart
@@ -1,6 +1,6 @@
 import 'package:irrigazione_iot/src/features/pumps/data/fake_pump_details_repository.dart';
 import 'package:irrigazione_iot/src/features/pumps/data/pump_repository.dart';
-import 'package:irrigazione_iot/src/features/pumps/domain/pump.dart';
+import 'package:irrigazione_iot/src/features/pumps/model/pump.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'pump_details_repository.g.dart';
