+++ lib/src/features/pumps/screen/add_pump/add_update_pump_form.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:irrigazione_iot/src/config/enums/form_types.dart';
-import 'package:irrigazione_iot/src/features/pumps/domain/pump.dart';
-import 'package:irrigazione_iot/src/features/pumps/presentation/add_pump/add_update_pump_controller.dart';
-import 'package:irrigazione_iot/src/features/pumps/presentation/add_pump/add_update_pump_form_content.dart';
+import 'package:irrigazione_iot/src/features/pumps/model/pump.dart';
+import 'package:irrigazione_iot/src/features/pumps/screen/add_pump/add_update_pump_controller.dart';
+import 'package:irrigazione_iot/src/features/pumps/screen/add_pump/add_update_pump_form_content.dart';
 import 'package:irrigazione_iot/src/utils/async_value_ui.dart';
 
 class AddUpdatePumpForm extends ConsumerWidget {
