+++ lib/src/features/pumps/screen/pump_status/pump_status_tile_wid_controller.dart
@@ -1,5 +1,5 @@
 import 'package:irrigazione_iot/src/features/pumps/data/pump_repository.dart';
-import 'package:irrigazione_iot/src/features/pumps/domain/pump.dart';
+import 'package:irrigazione_iot/src/features/pumps/model/pump.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'pump_status_tile_wid_controller.g.dart';
