+++ lib/src/features/sectors/domain/sector.dart
@@ -2,7 +2,7 @@
 import 'package:equatable/equatable.dart';
 
 import 'package:irrigazione_iot/src/config/enums/irrigation_enums.dart';
-import 'package:irrigazione_iot/src/features/pumps/domain/pump.dart';
+import 'package:irrigazione_iot/src/features/pumps/model/pump.dart';
 import 'package:irrigazione_iot/src/features/user_companies/domain/company.dart';
 
 typedef SectorID = String;
